move hijack to another file
